Refactored backend graphql queries
